+++ src/hooks/useTicketAddresses.ts
@@ -1,4 +1,4 @@
-import { parseEventsData } from "@/app/events/parseEventsData";
+import type { TicketInfoType } from "@/app/event/components/UserTicket";
 import type { Event } from "@/hooks/useCreateEventAttestation";
 import { CREATE_TICKET_SCHEMA_UID } from "@/hooks/useCreateTicketMultiAttestations";
 import { API_URL_MAPPING } from "@/lib/gqlEasAttestation";
@@ -6,7 +6,6 @@ import { GET_TICKETS_BY_EVENT_QUERY } from "@/lib/gqlEasAttestation/query";
 import { useMemo } from "react";
 import { useQuery } from "urql";
 import { useAccount, useChainId } from "wagmi";
-import type { TicketInfoType } from "../components/UserTicket";
 
 export const useTicketAddresses = ({
   eventId,
