+++ src/types/index.ts
@@ -1,0 +1,11 @@
+export interface DataEntry {
+  attester: string;
+  data: string;
+  decodedDataJson: string;
+  expirationTime: number;
+  id: string;
+  recipient: string;
+  refUID: string;
+  revocable: boolean;
+  revocationTime: number;
+}
